Move all tests to /test, fix errors, fix broken columnStateHelper-test test
