+++ testRunner.js
@@ -1,4 +1,4 @@
-const context = require.context('./src', true, /.+-test\.js$/)
+const context = require.context('./test', true, /.+-test\.js$/)
 context.keys().forEach(context)
 
 module.exports = context
